move pkg directory to api
