+++ app/dcr_api/biz/handler/job/BUILD.bazel
@@ -7,9 +7,9 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//app/dcr_api/biz/model/job",
+        "//app/dcr_api/biz/pkg/errno",
+        "//app/dcr_api/biz/pkg/utils",
         "//app/dcr_api/biz/service",
-        "//pkg/errno",
-        "//pkg/utils",
         "@com_github_cloudwego_hertz//pkg/app",
         "@com_github_cloudwego_hertz//pkg/common/hlog",
         "@com_github_cloudwego_hertz//pkg/protocol/consts",
