+++ app/dcr_api/biz/pkg/errno/BUILD.bazel
@@ -3,6 +3,6 @@ load("@rules_go//go:def.bzl", "go_library")
 go_library(
     name = "errno",
     srcs = ["errno.go"],
-    importpath = "github.com/manatee-project/manatee/pkg/errno",
+    importpath = "github.com/manatee-project/manatee/app/dcr_api/biz/pkg/errno",
     visibility = ["//visibility:public"],
 )
