+++ app/dcr_api/biz/service/BUILD.bazel
@@ -8,7 +8,7 @@ go_library(
     deps = [
         "//app/dcr_api/biz/dal/db",
         "//app/dcr_api/biz/model/job",
-        "//pkg/errno",
+        "//app/dcr_api/biz/pkg/errno",
         "@com_github_cloudwego_hertz//pkg/common/hlog",
         "@com_github_google_uuid//:uuid",
         "@com_github_pkg_errors//:errors",
