+++ app/dcr_api/biz/service/job_service.go
@@ -29,7 +29,7 @@ import (
 	"github.com/google/uuid"
 	"github.com/manatee-project/manatee/app/dcr_api/biz/dal/db"
 	"github.com/manatee-project/manatee/app/dcr_api/biz/model/job"
-	"github.com/manatee-project/manatee/pkg/errno"
+	"github.com/manatee-project/manatee/app/dcr_api/biz/pkg/errno"
 	"github.com/pkg/errors"
 )
